2023 day 10 part 2
yay learning math (:

After thinking about the problem a bit, I did realize that the solution was the interior area of a polygon, I just didn't know what the formula was so I looked it up.

Only thing I'm not personally happy with this my performance in this problem was hard coding the starting points. Finding S isn't hard, but determining which of the 4 compass directions to go in afterwards was some boiler plate I never got around to.

I am really, really happy with mapping the input to unicode to make everything readable